pyproject.toml
[project]
name = "ai-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gradio>=6.8.0",
    "ipython>=9.10.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "lxml>=6.0.2",
    "openai>=2.24.0",
    "playwright>=1.58.0",
    "pypdf2>=3.0.1",
    "pyppeteer>=0.0.25",
    "python-dotenv>=1.2.1",
    "sendgrid>=6.12.5",
]
